+++ plugins/users/src/components/page/User.tsx
@@ -15,7 +15,7 @@ import { useHistory, useParams } from 'react-router-dom';
 import React from 'react';
 import ReactMarkdown from 'react-markdown';
 
-import { IAuthProfile, getGravatarImageUrl } from '@kobsio/plugin-core';
+import { IUser, getGravatarImageUrl } from '@kobsio/plugin-core';
 import Teams from './Teams';
 
 interface IUserParams {
@@ -30,7 +30,7 @@ const User: React.FunctionComponent = () => {
   const history = useHistory();
   const params = useParams<IUserParams>();
 
-  const { isError, isLoading, error, data, refetch } = useQuery<IAuthProfile, Error>(
+  const { isError, isLoading, error, data, refetch } = useQuery<IUser, Error>(
     ['users/user', params.cluster, params.namespace, params.name],
     async () => {
       try {
@@ -72,7 +72,7 @@ const User: React.FunctionComponent = () => {
           <React.Fragment>
             <AlertActionLink onClick={(): void => history.push('/')}>Home</AlertActionLink>
             <AlertActionLink onClick={(): void => history.push('/users')}>Users</AlertActionLink>
-            <AlertActionLink onClick={(): Promise<QueryObserverResult<IAuthProfile, Error>> => refetch()}>
+            <AlertActionLink onClick={(): Promise<QueryObserverResult<IUser, Error>> => refetch()}>
               Retry
             </AlertActionLink>
           </React.Fragment>